test(images): cover EXIF orientation and face-crop geometry
images.rs was the last library file with a real testable gap: 72% line
coverage, with the orientation reader, the eight orientation transforms, and
the crop maths all unexercised. Now 86%.

Extracts apply_orientation as a pure function so all eight transforms can be
tested exhaustively without a fixture per case. The transform mapping is where
bugs actually live: 5 and 7 each combine a rotation with a flip and are easy
to transpose, so the tests assert exact pixel layouts on a deliberately
asymmetric 2x3 image. A square or symmetric fixture would pass for several
wrong mappings.

Orientation is read from a minimal JPEG built in the test: SOI, an APP1
segment holding a little-endian TIFF header whose IFD0 carries only the
Orientation tag, EOI. That avoids eight binary fixtures and avoids coupling
this crate's unit tests to crates/videre's fixture directory, and it makes the
malformed case (Orientation declared as LONG rather than SHORT, which must
fall back to 1) reachable at all.

Verified: all eight values round-trip, and a file declaring rotation is
actually rotated with its dimensions swapped, which the two halves tested
separately cannot show. Suite 512 passed, clippy back to 18.

+++ crates/videre-core/src/video_probe.rs
@@ -371,7 +371,7 @@ mod tests {
 
         // Claim the file is far larger than the bytes actually present, so the
         // bounds check passes and the cap is what rejects it.
-        let end = (MOOV_CAP + 1_000_000) as u64;
+        let end = MOOV_CAP + 1_000_000;
         assert!(has_video_track_in(&mut Cursor::new(f), end));
     }
